41 commits

+++ app/server/controllers/authController.js
@@ -11,7 +11,6 @@ initializePassport(
 )
 const checkAuth = async(req, res) => {
     try {
-        console.log('hello')
         res.status(200).json("hello")
     } catch (error) {
         res.status(503).json({ message: "not found" })
@@ -26,60 +25,52 @@ const token = async(req, res) => {
         const tokenSaved = tokens[0].refreshToken
         if (tokenSaved !== refreshToken) return res.status(403)
 
-        jwt.verify(refreshToken, process.env.REFRESH_TOKEN_SECRET, (error, user) => {
+        jwt.verify(refreshToken, process.env.REFRESH_TOKEN_SECRET, async(error, user) => {
             if (error) return res.status(403)
-            const accessToken = generateAccessToken({ name: user.name })
-
+            const accessToken = generateAccessToken({ id: user.id, name: user.name, email: user.email })
             res.json({ accessToken: accessToken })
         })
     } catch (error) {
         return res.status(500).json(error.message)
     }
-
 }
 const login = async(req, res) => {
-    // if (req.isAuthenticated()) {
-    //     console.log("user:" + req.user)
-    //     console.log(req.isAuthenticated())
-    //     res.status(200).json({ message: "authenticated" })
-    // } else {
-    //     res.status(503).send({ message: "not found" })
-    // }
     try {
         const { email: email, password: password } = req.body
-        console.log(email, password)
+        let crfToken
         if (!isValidEmail(email)) return res.Status(406).json({ message: "Provide correct email" })
 
         const checkEmail = await UserModel.findOne({ email: email })
         if (!checkEmail) return res.status(404).json({ message: "This email not exist" })
-
         bcrypt.compare(password, checkEmail.hash, async(error, result) => {
-            console.log(password, checkEmail.hash)
             if (error) {
                 return res.status(401).json({ message: "Wrong password" })
             }
             if (result) {
-                const user = { id: checkEmail._id, name: checkEmail.name, email: email }
+                const user = { id: checkEmail._id, name: checkEmail.name, email: email, admin: checkEmail.isAdmin }
                 const accessToken = generateAccessToken(user)
                 const refreshToken = jwt.sign(user, process.env.REFRESH_TOKEN_SECRET, { expiresIn: "1d" })
+                crfToken = refreshToken
                 const checkToken = await TokenModel.find()
-                if (checkToken.length !== 0) {
+                if (checkToken[0].refreshToken !== '') {
                     const idToUpdate = checkToken[0]._id
                     await TokenModel.findById(idToUpdate)
+                        .then(data => {
+                            if (!data) {
+                                return res.status(401).json({ message: "Unauthorized" })
+                            }
+                            data.refreshToken = refreshToken
+                            data.save().then(response => {
+                                if (response !== null) {
+                                    return res.json({ accessToken: accessToken, refreshToken: refreshToken, crfToken: crfToken })
+                                }
 
-                    .then(data => {
-                        if (!data) {
-                            return res.status(401).json({ message: "Unauthorized" })
-                        }
-                        data.refreshToken = refreshToken
-                        data.save().then(response => {
-                            return res.json({ accessToken: accessToken, refreshToken: refreshToken })
-                        }).catch(err => {
-                            res.status(401)
+                            }).catch(err => {
+                                res.status(401)
+                            })
+                        }).catch((err) => {
+                            res.status(401).json("Unauthorized")
                         })
-                    }).catch((err) => {
-                        res.status(401).json("Unauthorized")
-                    })
                 } else {
                     const newToken = new TokenModel({
                         refreshToken: refreshToken
@@ -111,7 +102,6 @@ const signup = async(req, res) => {
 
         const checkAccount = await UserModel.findOne({ email: email })
         if (checkAccount) {
-            console.log("already exists")
             return res.status(200).json({ message: "Email already exist" })
         }
         const user = new UserModel({
@@ -120,7 +110,6 @@ const signup = async(req, res) => {
             hash: hashPassword,
             isAdmin: false,
         })
-
         user.save().then((result) => {
             if (!result) {
                 return res.status(404).json({ message: "user not created" })
@@ -137,20 +126,21 @@ const recover = (req, res) => {
     res.send("recover")
 }
 const logout = async(req, res) => {
-    const { token: token } = req.body.token
+    const { token: token } = req.body
     try {
         const response = await TokenModel.find()
-        const getTokenSaved = response.refreshToken
+        if (!response) {
+            return res.sendStatus(401)
+        }
+        const getTokenSaved = response[0].refreshToken
         if (getTokenSaved !== token) return res.sendStatus(401)
         TokenModel.deleteOne({ refreshToken: getTokenSaved }).then(result => {
             res.sendStatus(204)
+
         }).catch(error => res.sendStatus(404).json(error))
 
     } catch (error) {
         res.sendStatus(500)
     }
 }
-
-
-
 module.exports = { token, checkAuth, login, signup, recover, logout }
